requirements.txt
rumps
py2app
psutil
jsontree
delegator.py